Gauge & Annotation scanner changes (#5)
Support both improved gauges where the gauge tracks in flight requests in methods (i.e. it increments when a method enters and decrements when a method exits).

Fixed how we scan for metric annotations to a pre-scan phase which enables us to make different optimisation to instrumentations - e.g. its now possible to add field level class information for metrics which we could not before as the method annotations were visited after the fields could be added.